Enable device top-ups
For Case 2, the battery is allowed to reach its maximum storage capacity (now defined by the `storage_capacity` parameter in the optimizer), when the `top_up` parameter in the optimizer is set to `True`. In this case the `soc_target` parameter is set equal to the `storage_capacity`. At the same time, the `device_energy_bounds` constraint has been relaxed to allow the SoC of the battery to surpass the `soc_max` only when it is necessary to avoid infeasibilities (such a case is tracked by the `supercharge_penalty` variable, which is penalized in the objective function).